bag => index
